Split feed controllers
