+++ routes/api.php
@@ -21,9 +21,11 @@ use App\Http\Controllers\Api\V1\Economy\PlanController;
 use App\Http\Controllers\Api\V1\Economy\ReceiptController;
 use App\Http\Controllers\Api\V1\Economy\SubscriptionController;
 use App\Http\Controllers\Api\V1\Economy\TransactionController;
-use App\Http\Controllers\Api\V1\Feeds\CasinoFloorController;
+use App\Http\Controllers\Api\V1\Feeds\AchievementFeedController;
+use App\Http\Controllers\Api\V1\Feeds\ChallengeFeedController;
 use App\Http\Controllers\Api\V1\Feeds\LeaderboardController;
 use App\Http\Controllers\Api\V1\Feeds\LiveScoresController;
+use App\Http\Controllers\Api\V1\Feeds\TournamentFeedController;
 use App\Http\Controllers\Api\V1\Games\ActionController;
 use App\Http\Controllers\Api\V1\Games\GameAbandonController;
 use App\Http\Controllers\Api\V1\Games\GameConcedeController;
@@ -178,9 +180,9 @@ Route::prefix('v1')->group(function () {
         Route::get('/games', [LiveScoresController::class, 'games']);
         Route::get('/wins', [LiveScoresController::class, 'wins']);
         Route::get('/leaderboards/{gameTitle}', [LeaderboardController::class, 'show']);
-        Route::get('/tournaments', [CasinoFloorController::class, 'tournaments']);
-        Route::get('/challenges', [CasinoFloorController::class, 'challenges']);
-        Route::get('/achievements', [CasinoFloorController::class, 'achievements']);
+        Route::get('/tournaments', [TournamentFeedController::class, 'show']);
+        Route::get('/challenges', [ChallengeFeedController::class, 'show']);
+        Route::get('/achievements', [AchievementFeedController::class, 'show']);
     });
 
     // ========================================
